Added some support for transforms
When doing this first, I was transforming the objects, I need to go back and transform the rays instead at least for spheres

+++ src/Renderer/FragmentOutput.h
@@ -14,7 +14,7 @@ public:
 		for (Fragment & frag : fragBuffer){
 			color += frag.getColor();
 		}
-		color /= glm::vec3(fragBuffer.size());
+		color /= glm::vec3((float) fragBuffer.size());
 	}
 
 	glm::vec3 getColor() { return color; }
